requirements.txt
cryptography==46.0.3
mysql_connector_repackaged==0.3.1
pandas==2.3.3
Pillow==12.0.0
pycryptodome==3.23.0
streamlit==1.51.0
~treamlit==1.51.0
